Extract general elements from HubClientConnector into SocketClientConnector
